requirements.txt
Flask~=1.1
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
Flask==2.1.1
Flask-MySQLdb==1.0.1
fonttools==4.31.2
idna==3.3
importlib-metadata==4.11.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
ks-api-client==1.1.0
MarkupSafe==2.1.1
matplotlib==3.5.1
mysqlclient==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pip==22.1.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.3.0
pytz==2022.1
requests==2.26.0
setuptools==57.4.0
six==1.16.0
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.1
zipp==3.8.0
mysql-connector-python==8.0.29
protobuf==4.21.1
